move `TxInclusionProof` into queryable.rs
