+++ sequencer/src/block2.rs
@@ -1,4 +1,3 @@
-use crate::block2::tx_iterator::TxIndex;
 use crate::{BlockBuildingSnafu, Transaction, VmId};
 use ark_bls12_381::Bls12_381;
 use commit::{Commitment, Committable};
@@ -12,7 +11,7 @@ use jf_primitives::{
             payload_prover::{LargeRangeProof, SmallRangeProof},
             Advz,
         },
-        payload_prover::{PayloadProver, Statement},
+        payload_prover::PayloadProver,
     },
 };
 use serde::{Deserialize, Serialize};
@@ -463,135 +462,13 @@ pub fn get_ns_payload_range(
     start..end
 }
 
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-pub struct TxInclusionProof {
-    tx_table_len: TxTableEntry,
-    tx_table_len_proof: RangeProof,
-
-    tx_table_range_start: Option<TxTableEntry>, // `None` for the 0th tx
-    tx_table_range_end: TxTableEntry,
-    tx_table_range_proof: RangeProof,
-
-    tx_payload_proof: Option<RangeProof>, // `None` if the tx has zero length
-}
-
-impl TxInclusionProof {
-    // TODO currently broken, fix in https://github.com/EspressoSystems/espresso-sequencer/issues/1010
-    //
-    // - We need to decide where to store VID params.
-    // - Returns `None` if an error occurred.
-    // - Use of `Result<(),()>` pattern to enable use of `?` for concise abort-on-failure.
-    #[allow(dead_code)] // TODO temporary
-    #[allow(clippy::too_many_arguments)]
-    fn verify<V>(
-        &self,
-        tx: &Transaction,
-        tx_index: TxIndex,
-        vid: &V,
-        vid_commit: &V::Commit,
-        vid_common: &V::Common,
-    ) -> Option<Result<(), ()>>
-    where
-        V: PayloadProver<RangeProof>,
-    {
-        V::is_consistent(vid_commit, vid_common).ok()?;
-
-        // Verify proof for tx payload.
-        // Proof is `None` if and only if tx has zero length.
-        let tx_payload_range = tx_payload_range(
-            &self.tx_table_range_start,
-            &self.tx_table_range_end,
-            &self.tx_table_len,
-            V::get_payload_byte_len(vid_common),
-        )?;
-        match &self.tx_payload_proof {
-            Some(tx_payload_proof) => {
-                if vid
-                    .payload_verify(
-                        Statement {
-                            payload_subslice: tx.payload(),
-                            range: tx_payload_range,
-                            commit: vid_commit,
-                            common: vid_common,
-                        },
-                        tx_payload_proof,
-                    )
-                    .ok()?
-                    .is_err()
-                {
-                    return Some(Err(())); // TODO it would be nice to use ? here...
-                }
-            }
-            None => {
-                if !tx.payload().is_empty() || !tx_payload_range.is_empty() {
-                    return None; // error: nonempty payload but no proof
-                }
-            }
-        };
-
-        // Verify proof for tx table len.
-        if vid
-            .payload_verify(
-                Statement {
-                    payload_subslice: &self.tx_table_len.to_bytes(),
-                    range: 0..TxTableEntry::byte_len(),
-                    commit: vid_commit,
-                    common: vid_common,
-                },
-                &self.tx_table_len_proof,
-            )
-            .ok()?
-            .is_err()
-        {
-            return Some(Err(()));
-        }
-
-        // Verify proof for tx table entries.
-        // Start index missing for the 0th tx
-        let index: usize = tx_index.tx_idx; // TODO fix in https://github.com/EspressoSystems/espresso-sequencer/issues/1010
-        let mut tx_table_range_bytes =
-            Vec::with_capacity(2usize.checked_mul(TxTableEntry::byte_len())?);
-        let start = if let Some(tx_table_range_start) = &self.tx_table_range_start {
-            if index == 0 {
-                return None; // error: first tx should have empty start index
-            }
-            tx_table_range_bytes.extend(tx_table_range_start.to_bytes());
-            index * TxTableEntry::byte_len()
-        } else {
-            if index != 0 {
-                return None; // error: only the first tx should have empty start index
-            }
-            TxTableEntry::byte_len()
-        };
-        tx_table_range_bytes.extend(self.tx_table_range_end.to_bytes());
-        let range = start
-            ..index
-                .checked_add(2)?
-                .checked_mul(TxTableEntry::byte_len())?;
-
-        if vid
-            .payload_verify(
-                Statement {
-                    payload_subslice: &tx_table_range_bytes,
-                    range,
-                    commit: vid_commit,
-                    common: vid_common,
-                },
-                &self.tx_table_range_proof,
-            )
-            .ok()?
-            .is_err()
-        {
-            return Some(Err(()));
-        }
-
-        Some(Ok(()))
-    }
-}
-
 #[cfg(test)]
 mod test {
-    use super::{test_vid_factory, Payload, Transaction, TxInclusionProof, TxIndex, TxTableEntry};
+    use super::{test_vid_factory, Payload, Transaction, TxTableEntry};
+    use crate::block2::{
+        queryable::{self},
+        tx_iterator::TxIndex,
+    };
     use async_compatibility_layer::logging::{setup_backtrace, setup_logging};
     use helpers::*;
     use hotshot_query_service::availability::QueryablePayload;
@@ -981,14 +858,11 @@ mod test {
 
         // make a fake proof for a nonexistent tx in the small block
         let tx = Transaction::new(crate::VmId(0), Vec::new());
-        let proof = TxInclusionProof {
-            tx_table_len: block.get_tx_table_len(),
-            tx_table_len_proof: block.get_tx_table_len_proof(&vid).unwrap().clone(),
-            tx_table_range_start: None,
-            tx_table_range_end: TxTableEntry::from_usize(1),
-            tx_table_range_proof: vid.payload_proof(&block.payload, 0..3).unwrap(),
-            tx_payload_proof: None,
-        };
+        let proof = queryable::gen_tx_proof_for_testing(
+            block.get_tx_table_len(),
+            block.get_tx_table_len_proof(&vid).unwrap().clone(),
+            vid.payload_proof(&block.payload, 0..3).unwrap(),
+        );
 
         // test: fake proof should get rejected
         // TODO should return Some(Err()) instead of None
